changed hubToleranceDegrees variable to double

+++ src/main/java/frc/robot/Constants.java
@@ -188,7 +188,7 @@ public final class Constants {
   public final class Flywheel {
     public static final int flywheelMotorA = 40;
     public static final int flywheelMotorB = 41;
-    public static final int hubToleranceDegrees = 3;
+    public static final double hubToleranceDegrees = 3;
 
     public static final int encoderUnitsPerRotation = 2048;
